WA-234 fetch safe info regularly in safe view
