+++ src/routes/safe/store/actions/updateSafe.js
@@ -1,0 +1,8 @@
+// @flow
+import { createAction } from 'redux-actions'
+
+export const UPDATE_SAFE = 'UPDATE_SAFE'
+
+const updateSafe = createAction(UPDATE_SAFE)
+
+export default updateSafe
